+++ contracts/interfaces/bridge/IBridgeManager.sol
@@ -54,7 +54,7 @@ interface IBridgeManager is IBridgeManagerEvents {
   function getFullBridgeOperatorInfos()
     external
     view
-    returns (address[] memory governors, address[] memory bridgeOperators, uint256[] memory weights);
+    returns (address[] memory governors, address[] memory bridgeOperators, uint96[] memory weights);
 
   /**
    * @dev Returns total weights of the governor list.
@@ -92,19 +92,19 @@ interface IBridgeManager is IBridgeManagerEvents {
    * @param governor The address of the governor to get the vote weight for.
    * @return voteWeight The vote weight of the specified governor.
    */
-  function getGovernorWeight(address governor) external view returns (uint256);
+  function getGovernorWeight(address governor) external view returns (uint96);
 
   /**
    * @dev External function to retrieve the vote weight of a specific bridge operator.
    * @param bridgeOperator The address of the bridge operator to get the vote weight for.
    * @return weight The vote weight of the specified bridge operator.
    */
-  function getBridgeOperatorWeight(address bridgeOperator) external view returns (uint256 weight);
+  function getBridgeOperatorWeight(address bridgeOperator) external view returns (uint96 weight);
 
   /**
    * @dev Returns the weights of a list of governor addresses.
    */
-  function getGovernorWeights(address[] calldata governors) external view returns (uint256[] memory weights);
+  function getGovernorWeights(address[] calldata governors) external view returns (uint96[] memory weights);
 
   /**
    * @dev Returns an array of all governors.
